(add) upload & latex
