+++ packages/api-server/src/admin.html
@@ -1,9 +1,28 @@
 <head>
-  <link href='https://fonts.googleapis.com/css?family=Poppins' rel='stylesheet'>
+  <script>
+    window.MathJax = {
+      tex: {
+        loader: {load: ['[tex]/textmacros']},
+        inlineMath: [
+          ['$', '$'],
+          ['\\(', '\\)']
+        ],
+        displayMath: [             // start/end delimiter pairs for display math
+          ['$$', '$$'],
+          ['\\[', '\\]']
+        ],
+        processEscapes: true,      // use \$ to produce a literal dollar sign
+        processEnvironments: true, // process \begin{xxx}...\end{xxx} outside math mode
+        processRefs: true,         // process \ref{...} outside of math mode
+        packages: {'[+]': ['textmacros']},
+      }
+    };
+  </script>
+  <script id="MathJax-script" async src="https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js"></script>
   <style> 
     body{
       margin:40px auto;max-width:700px;line-height:1.6;font-size:18px;color:#444;padding:0 10px;
-      font-family: 'Poppins'
+      font-family: 'Ariel'
     }
     h1,h2,h3{line-height:1.2}
     input,div,button { margin: 0.2rem }
@@ -83,7 +102,8 @@
 </div>
 
 <h2>Problems</h2>
-<button id='problem-polygon'>polygon add</button> (zipped package)
+<button id='problem-polygon'>add</button>
+<button id='problem-polygon-replace'>replace</button> (zipped package)
 <input type="file" id="problem-polygon-input"> <br>
 <button id='problem-get'>get</button> 
 <button id='problem-delete'>delete</button> <br>
@@ -150,6 +170,48 @@
         alert(msg);
         throw err
       })
+
+
+  let typesetting = undefined;
+  const trymathjax = async _ => {
+    MathJax.typeset();
+    typesetting = false;
+  }
+
+  const mathjax = _ => {
+    if (!typesetting)
+      trymathjax()
+  }
+
+  const formatProblem = problem => 
+  {
+    const {name, constraints, context, inputFormat, outputFormat} = problem;
+    let html = `
+    <div style='text-align: center'>
+    <h3>${name}</h3>
+    time limit / test: ${constraints.time} ms <br>
+    memory limit / test: ${constraints.memory} MB 
+    </div>
+    ${context}<br>
+    <b>Input</b><br>
+    ${inputFormat}<br>
+    <b>Output</b><br>
+    ${outputFormat}<br>`;
+
+    html = html
+      .replaceAll("\\begin{center}", "<p style='text: center'>")
+      .replaceAll("\\end{center}", "</p>")
+      .replaceAll("\\begin{enumerate}", "<ol>")
+      .replaceAll("\\item", "<li>")
+      .replaceAll("\\end{enumerate}", "</ol>")
+      .replaceAll("\\begin{itemize}", "<ul>")
+      .replaceAll("\\item", "<li>")
+      .replaceAll("\\end{itemize}", "</ul>")
+      .replaceAll(/\\t{(.*?)}/g, "<code>$1</code>")
+      .replaceAll(/\\bf{(.*?)}/g, "<b>$1</b>")
+
+    return html;
+  }
 
   document.getElementById('signup').onclick = async _ => {
 
@@ -310,6 +372,28 @@
       .then(res => alert('package uploaded'))
   }
 
+  document.getElementById('problem-polygon-replace').onclick = async _ => {
+
+    console.log('problem-polygon-replace');
+
+    if (getDomainId() === undefined) return alert('no domain selected');
+    if (getProblemId() === undefined) return alert('no problem selected');
+    
+    if (!document.getElementById('problem-polygon-input').files[0])
+      return alert('no polygon file selected');
+
+    const { uploadId } = await betterFetch(`/v1/domains/${getDomainId()}/problems/${getProblemId()}/polygon-upload-session`);
+
+    const formData = new FormData();
+    formData.append('package', document.getElementById('problem-polygon-input').files[0]);
+
+    const res = await betterFetch(uploadURL + `/polygon/${uploadId}`, {
+      method: "POST",
+      body: formData 
+    })
+      .then(res => alert('package uploaded'))
+  }
+
   document.getElementById('problem-get').onclick = async _ => {
 
     console.log('problem-get');
@@ -319,26 +403,25 @@
     const problems = await betterFetch(`/v1/domains/${getDomainId()}/problems/`)
 
     let html = '';
-    for (const { name, id, context, inputFormat, outputFormat, constraints } of problems) {
+    for (const problem of problems) {
 
-      const problemHTML = `
-      ${constraints.time}ms <br>
-      ${constraints.memory}MB <br>
-      <hr>
-      ${context}<br>
-      ${inputFormat}<br>
-      ${outputFormat}<br>
-      <hr>`;
+      let problemHTML = formatProblem(problem);
+
+      const {id, name} = problem;
       html += ` 
       <details>
         <summary>
           <input type='radio' name='problem-radio' value='${id}'>
           <label for='${id}'><b>${name}</b> (<i>${id}</i>)</label>
         </summary>
-        ${problemHTML}
+        <p>
+          ${problemHTML}
+        </p>
       </details>`
     }
-   document.getElementById('problem-list').innerHTML = html;
+    document.getElementById('problem-list').innerHTML = html;
+
+    mathjax()
   }
 
   document.getElementById('series-add').onclick = async _ => {
